requirements.txt
numpy
opencv-python
Pillow
facenet-pytorch
scikit-learn
joblib